+++ component/include/pl_modbus_server.h
@@ -15,6 +15,8 @@ public:
   static const std::string defaultName; 
   /// @brief Default read operation timeout in FreeRTOS ticks
   static constexpr TickType_t defaultReadTimeout = 2;
+  /// @brief Default write operation timeout in FreeRTOS ticks
+  static constexpr TickType_t defaultWriteTimeout = 300 / portTICK_PERIOD_MS;
 
   /// @brief Creates a stream Modbus server with shared transaction buffer
   /// @param stream stream